Remove ancient comments
For those not familiar with this thing's history, it originally
started off as a simple Ogg Vorbis music player for Cave Story.

Over time it just gained more and more features, and was refactored
to the point that it became a general-purpose framework. That's when
I decided to break it off into its own project.