Big disgusting decoder refactor part 1
This is a nightmare.

Alright, basically, I want to make clownaudio a suitable backend for
sequencers, so I need to rethink how sample conversion works: the
user should be able to override the source sample rate before the
resampler converts it to the output format. This commit is just
preperation-work for that.