Allow sounds to exist after they've finished
The fire-and-forget philosophy has worked well so far, but sometimes
you *want* to keep a sound around after it's finished. For example,
combined with the newly-added RewindSound function, you can play the
same sound multiple times, without constantly creating new instances.
Of course, the downside to that approach is that 'playing' the same
sound twice will cause the second one to overwrite the first, but
this might be desireable.

In my case, I want this because it's the behaviour Cave Story wants.