Re-enable dr_flac's Ogg-encapsulated FLAC support
Previously, this was disabled because all files whose filename ended
with '.ogg' were redirected to the Ogg Vorbis decoders - dr_flac
would never get a chance to decode them. But now that decoder
selection has been overhauled, this is no longer a problem.

It would be nice if I had a test .ogg file to test this with...