Killed-off the Tremor decoder backend
This thing kind of sucks. Not only does it corrupt the stack if it
tries to load a FLAC file, but it also does it when loading an Opus
file. Sanity checks? Hello?

Even the specialised 'ov_test' functions don't work. You know, the
functions specifically for testing if a file is a Vorbis file?

There's no point to supporting an integer-only Vorbis decoder anyway,
when clownaudio itself uses floats extensively.