Mark libxmp-lite as an object library
Trying to fix an annoying error, where if you create a static
clownaudio lib in MSYS2, which uses the internal copy of libxmp-lite,
and then install it, pkg-config won't link libxmp-lite.

Apparently, CMake has a thing called an object library, which is a
library that never gets linked independently - it becomes part of the
parent library. Just what I need, right?