Begin adding framework for requesting formats
The idea is to once again unify the decoder API, just like in the
v0.1 days. This will allow me to make the 'decoder' pipeline
configurable again.

This has the added benefit of laying the basic groundwork for
allowing decoders to match the sample rate of the mixer (once dynamic
resampling is made optional).