Improve usability
Now, the clownaudio headers are just as visible to clownaudio itself
as they are to the user. This allows 'unity builds' like the
Makefile.old one to compile without needing to special-case for user
sources and clownaudio sources.

Also, I added a new way to work around the lack of
`clownaudio_export.h` in unity builds: as part of your build system,
defined `CLOWNAUDIO_EXPORT` and `CLOWNAUDIO_NO_EXPORT` as nothing.
This will prevent the headers from trying to include
`clownaudio_export.h`.