Fix resampling
It would always assume the input buffer was filled completely, even
when it's actually only filled partially, causing it to offset
incorrectly and read garbage data.

This is most noticable if a sound fully decodes to a size smaller
than the input buffer, in which case the entire sound will be
audibly incorrect.