Add a linked-list for currently-playing sounds
This is another optimisation, this time for avoiding the need for the
mixer to iterate over every sound, including the ones that aren't
playing.

By having a linked list that only includes sounds that are playing,
the mixer does as little work as possible. Of course, this comes at a
slight memory penalty as every sound now has two more pointers.