Replace the miniaudio resampler with a custom one
With this, the hard dependency on miniaudio is no more.

This new resampler is still very experimental, so I'd like to keep
the miniaudio resampler around for a bit longer.

Anyway, this new resampler is clownresampler: a windowed-sinc filter
implemented as a single-header library written in C89. By integrating
it into clownaudio, I hope to test it in more varied environments.