Add oswrapper_audio backend
oswrapper_audio is basically a thin wrapper over OS provided audio decoding functionality on supported platforms. Currently, it only has an implementation for macOS, but I plan to add a Windows implementation once I figure out how to use the Microsoft Media Foundation APIs. I was originally working on a macOS specific AudioToolbox decoder backend, but I decided to spin it off into a library, so the API is very similar to a clownaudio decoder.